+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/exceptions.scala
@@ -32,3 +32,8 @@ case class TableException(msg: String) extends RuntimeException(msg)
   * Exception for all errors occurring during validation phase.
   */
 case class ValidationException(msg: String) extends RuntimeException(msg)
+
+/**
+  * Exception for unwanted method calling on unresolved expression.
+  */
+case class UnresolvedException(msg: String) extends RuntimeException(msg)
