+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/expressions/cast.scala
@@ -20,9 +20,8 @@ package org.apache.flink.api.table.expressions
 import org.apache.calcite.rex.RexNode
 import org.apache.calcite.tools.RelBuilder
 
-import org.apache.flink.api.common.typeinfo.BasicTypeInfo._
-import org.apache.flink.api.common.typeinfo.{NumericTypeInfo, TypeInformation}
-import org.apache.flink.api.table.typeutils.TypeConverter
+import org.apache.flink.api.common.typeinfo.TypeInformation
+import org.apache.flink.api.table.typeutils.{TypeCoercion, TypeConverter}
 import org.apache.flink.api.table.validate._
 
 case class Cast(child: Expression, resultType: TypeInformation[_]) extends UnaryExpression {
@@ -39,36 +38,10 @@ case class Cast(child: Expression, resultType: TypeInformation[_]) extends Unary
   }
 
   override def validateInput(): ExprValidationResult = {
-    if (Cast.canCast(child.resultType, resultType)) {
+    if (TypeCoercion.canCast(child.resultType, resultType)) {
       ValidationSuccess
     } else {
       ValidationFailure(s"Unsupported cast from ${child.resultType} to $resultType")
     }
-  }
-}
-
-object Cast {
-
-  /**
-    * all the supported cast type
-    */
-  def canCast(from: TypeInformation[_], to: TypeInformation[_]): Boolean = (from, to) match {
-    case (from, to) if from == to => true
-
-    case (_, STRING_TYPE_INFO) => true
-
-    case (_, DATE_TYPE_INFO) => false // Date type not supported yet.
-    case (_, VOID_TYPE_INFO) => false // Void type not supported
-    case (_, CHAR_TYPE_INFO) => false // Character type not supported.
-
-    case (STRING_TYPE_INFO, _: NumericTypeInfo[_]) => true
-    case (STRING_TYPE_INFO, BOOLEAN_TYPE_INFO) => true
-
-    case (BOOLEAN_TYPE_INFO, _: NumericTypeInfo[_]) => true
-    case (_: NumericTypeInfo[_], BOOLEAN_TYPE_INFO) => true
-
-    case (_: NumericTypeInfo[_], _: NumericTypeInfo[_]) => true
-
-    case _ => false
   }
 }
